requirements.txt
aiogram>=3.4,<4
SQLAlchemy>=2.0.30
aiosqlite>=0.19
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.29
